pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "transparent-company"
version = "0.1.0"
description = "Transparent Company MVP+ with verifiable disclosure, anchoring, and BI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115,<0.116",
  "uvicorn[standard]>=0.30,<0.31",
  "pydantic>=2.8,<2.9",
  "pydantic-settings>=2.3,<2.4",
  "sqlalchemy>=2.0.32,<2.1",
  "psycopg2-binary>=2.9.9,<2.10",
  "pandas>=2.2.2,<2.3",
  "PyNaCl>=1.5.0,<1.6",
  "beancount>=2.3.6,<2.4",
  "eventsourcing>=9.3.0,<9.4",
  "minio>=7.2.7,<7.3",
  "requests>=2.32.3,<2.33",
  "immudb-py>=1.5.0,<1.6",
  "python-dateutil>=2.9.0,<2.10",
  "orjson>=3.10.7,<3.11"
]

[project.optional-dependencies]
test = [
  "pytest>=8.3.2,<8.4",
  "pytest-cov>=5.0.0,<5.1",
  "httpx>=0.27.0,<0.28"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.setuptools.package-data]
"app.dashboard.superset" = ["*.py", "*.sh", "*.sql"]
